pyproject.toml
[project]
name = "agentic_blog_writer"
version = "0.1.0"
description = "Agentic blog writer"
authors = []
dependencies = [
    "openai>=1.0.0,<2.0.0",
    "python-dotenv>=1.0.0",
    "python-slugify>=8.0.0,<9.0.0", # For creating filesystem-friendly names
    "fastapi",
    "uvicorn",
    "pydantic",
    "rich"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.0,<8.0.0",
    "ruff"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0,<8.0.0",
    "ruff"
] 